Added Core mod
Fixed IC2 enetgy API without interrupting IC2's functional & their
addons

Closes #53

+++ src/main/java/gregtechmod/api/interfaces/IBasicEnergyContainer.java
@@ -50,7 +50,7 @@ public interface IBasicEnergyContainer {
 	/**
 	 * Inject Energy Call for Electricity.
 	 */
-	public boolean injectEnergyUnits(byte aSide, int aVoltage, int aAmperage);
+	public boolean injectEnergyUnits(byte aSide, int aVoltage, double aAmperage);
 	
 	/**
 	 * Drain Energy Call for Electricity.
